requirements.txt
flask
gunicorn
keras==2.8.0
pillow==6.2.2
protobuf==3.20.*
tensorflow==2.8.0